pyproject.toml
[build-system]
requires = ["poetry==0.12.17"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "morattendance_standalone"
description = "MorAttendance"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.7.3"
django = "^2.2"
django-split-settings = "^0.3"
django-axes = "^5.0"
django-csp = "^3.5"
django-health-check = "^3.10"
django-feature-policy = "^2.3"
psycopg2 = "^2.8"
gunicorn = "^19.9"
python-decouple = "^3.1"
dump-env = "^1.0"
bcrypt = "^3.1"
django_redis = "^4.10"
django-redis = "^4.10"
djangorestframework = "^3.10"
django_extensions = "^2.2"
django-extensions = "^2.2"
nfcpy = "^1.0"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.0"
django-querycount = "^0.7"
nplusone = "^1.0"
wemake-python-styleguide = "^0.10"
flake8-pytest = "^1.3"
flake8-django = "^0.0.4"
pytest = "^5.0"
pytest-django = "^3.5"
pytest-cov = "^2.7"
pytest-randomly = "^3.0"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^0.9"
pytest-timeout = "^1.3"
pre-commit = "^1.17"
xenon = "^0.5"
mypy = "^0.720"
ipython = "^7.7"
sphinx = "^2.1"
sphinx-autodoc-typehints = "^1.7"
doc8 = "^0.8"
yamllint = "^1.15"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"
